SDK update: API changes, fixes #1

+++ overlays/Nuti.framework/Versions/A/Headers/renderers/LineRenderer.h
@@ -26,7 +26,7 @@ public:
                                     GLuint a_texCoord,
                                     std::vector<unsigned char>& colorBuf,
                                     std::vector<float>& coordBuf,
-                                    std::vector<unsigned int>& indexBuf,
+                                    std::vector<unsigned short>& indexBuf,
                                     std::vector<float>& texCoordBuf,
                                     std::vector<const LineDrawData*>& drawDataBuffer,
                                     LRUTextureCache<std::shared_ptr<Bitmap> >& styleCache,
@@ -59,7 +59,7 @@ private:
 
     std::vector<unsigned char> _colorBuf;
 	std::vector<float> _coordBuf;
-    std::vector<unsigned int> _indexBuf;
+    std::vector<unsigned short> _indexBuf;
 	std::vector<float> _texCoordBuf;
 
 	const Shader* _shader;
