+++ overlays/Nuti.framework/Versions/A/Headers/renderers/PointRenderer.h
@@ -43,7 +43,7 @@ private:
                                       GLuint a_texCoord,
                                       std::vector<unsigned char>& colorBuf,
                                       std::vector<float>& coordBuf,
-                                      std::vector<unsigned int>& indexBuf,
+                                      std::vector<unsigned short>& indexBuf,
                                       std::vector<float>& texCoordBuf,
                                       std::vector<const PointDrawData*>& drawDataBuffer,
                                       const MapBounds& texCoordBounds,
@@ -60,7 +60,7 @@ private:
 
     std::vector<unsigned char> _colorBuf;
 	std::vector<float> _coordBuf;
-    std::vector<unsigned int> _indexBuf;
+    std::vector<unsigned short> _indexBuf;
 	std::vector<float> _texCoordBuf;
 
 	const Shader* _shader;
